requirements.txt
source venv/bin/activate
azure-functions
psycopg2-binary
azure-functions
azure-storage-blob
matplotlib
numpy
psycopg2-binary
azure-iot-hub
azure-iot-device
python-dotenv
